reimplement: SHC_3BB0A8C1_0x00492C60 100.00% Reimplemented

+++ src/OpenSHC/UI/MenuItems/SaveLoadMap/MenuItemRenderFunction_SaveLoadMap_Scrollbar.cpp
@@ -19,8 +19,8 @@ namespace UI {
                 int param_1, int thumbYPos, int param_3, int thumbHeight, BOOLEnum isDragged)
             {
                 MACRO_CALL_MEMBER(OpenSHC::UI::Rendering::PencilRenderCore_Func::drawScrollbar,
-                    DAT_PencilRenderCore::ptr)(DAT_ButtonX::instance, (int)((int)(DAT_ButtonY::instance)),
-                    (int)((int)(DAT_ButtonH::instance)), thumbYPos, isDragged, thumbHeight, 0);
+                    DAT_PencilRenderCore::ptr)(DAT_ButtonX::instance, DAT_ButtonY::instance, DAT_ButtonH::instance,
+                    thumbYPos, isDragged, thumbHeight, 0);
             }
 
         }
